Fix error when printing out non-URL destinations
The non-ANSI report was trying to print out URI of the destination. That makes sense most of the time but not when the ‘destination’ is, say, a data URL.

This commit also improves error reporting in case we hit the same problem again.

This should fix https://github.com/filiph/linkcheck/issues/1.

Bump version to 0.2.13.