Show details for redirected links when --show-redirects is set
PR #54 added a `--show-redirects` flag! But although it adds a redirects count
to the summary, it won't show the details about WHICH links were redirected
unless there was *also* at least one error or warning. So if you want to
actually update your redirected links, you need to temporarily break an
unrelated link first. 😆

This commit adds "they requested redirects, and there was at least one redirect"
to the list of conditions that can trigger a full report.